Use WideWine UUID and random UUID to replace IMEI in registration
TelephonyManager.getDeviceId() is disabled since Android 10 unless the
devices holds a READ_PRIVILEGED_PHONE_STATE permission which is only
allowed for system applications. To fix the registration on these
devices, the device ID is replaced with a WideWine UUID or if that is
not available, a random UUID.

The random UUID will also be included in the SHA-512 hash to avoid the
scenario where all other fields but timestamp would fall back to their
default value.

See more:
https://developer.android.com/about/versions/10/privacy/changes